add for benchmark

+++ src/function_transformer_attention.py
@@ -193,7 +193,7 @@ class SpGraphTransAttentionLayer(nn.Module):
     if not self.opt['beltrami'] and self.opt['attention_type'] == "exp_kernel":
       prods = self.output_var ** 2 * torch.exp(-(torch.sum((src - dst_k) ** 2, dim=1) / (2 * self.lengthscale ** 2)))
     elif self.opt['attention_type'] == "scaled_dot":
-      prods = torch.sum(src * dst_k, dim=1) / np.sqrt(self.d_k)
+      prods = torch.sum(src * dst_k, dim=1, dtype=src.dtype) / np.sqrt(self.d_k)
     elif self.opt['attention_type'] == "cosine_sim":
       cos = torch.nn.CosineSimilarity(dim=1, eps=1e-5)
       prods = cos(src, dst_k)
